requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
Pillow==11.1.0
cryptography==44.0.0
MarkupSafe==3.0.2
markupsafe==3.0.2
requests==2.32.3
PyYAML
ldap3==2.9.1
lxml==5.3.0